+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - [SpinElement](src/spinElement.ts). Wrong render on option `fit` on Safari 14-
 - [PopupElement](src/popup/popupElement.ts). Animation for opacity doesn't work on Safari 14-
 - [PopupElement](src/popup/popupElement.ts). Animation affects on bluring text if scroll body during the animation
+- [PopupElement](src/popup/popupElement.ts). Allow to open/close on double-click
 
 - **New**: